chore: move cover.py to the top-level module and remove match_utils
